backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
django-cors-headers
drf-nested-routers
django-import-export
django-jazzmin
Faker
gunicorn
dj-database-url
whitenoise
